Move snapshot test
