+++ hatchery/tests/boxen.rs
@@ -5,7 +5,7 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 use dallo::ModuleId;
-use hatchery::{module_bytecode, Error, Receipt, SnapshotId, World};
+use hatchery::{module_bytecode, Error, Receipt, World};
 use std::path::PathBuf;
 
 #[test]
@@ -52,48 +52,5 @@ pub fn box_set_store_restore_get() -> Result<(), Error> {
 
     assert_eq!(*value, Some(0x23));
 
-    Ok(())
-}
-
-#[test]
-pub fn world_snapshot_persist_restore() -> Result<(), Error> {
-    let mut world = World::ephemeral()?;
-    let id = world.deploy(module_bytecode!("box"))?;
-
-    fn create_snapshot(
-        world: &mut World,
-        id: ModuleId,
-        arg: i16,
-    ) -> Result<SnapshotId, Error> {
-        let _: Receipt<()> = world.transact(id, "set", arg)?;
-
-        let value = world.query::<_, Option<i16>>(id, "get", ())?;
-
-        assert_eq!(*value, Some(arg));
-
-        world.persist()
-    }
-
-    fn restore_snapshot(
-        world: &mut World,
-        id: ModuleId,
-        snapshot_id: &SnapshotId,
-        arg: i16,
-    ) -> Result<(), Error> {
-        world.restore(&snapshot_id)?;
-        let value = world.query::<_, Option<i16>>(id, "get", ())?;
-        let _: Receipt<()> = world.transact(id, "mem_snap", ())?;
-        assert_eq!(*value, Some(arg));
-        Ok(())
-    }
-
-    let mut snapshot_ids = Vec::new();
-    let random_i = vec![3, 1, 0, 4, 2];
-    for i in 0..random_i.len() {
-        snapshot_ids.push(create_snapshot(&mut world, id, i as i16)?);
-    }
-    for i in random_i {
-        restore_snapshot(&mut world, id, &snapshot_ids[i], (i) as i16)?;
-    }
     Ok(())
 }
