+++ hatchery/tests/snapshot.rs
@@ -4,7 +4,51 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-use hatchery::{module_bytecode, Error, Receipt, World};
+use dallo::ModuleId;
+use hatchery::{module_bytecode, Error, Receipt, SnapshotId, World};
+
+#[test]
+pub fn snapshot_persist_restore() -> Result<(), Error> {
+    let mut world = World::ephemeral()?;
+    let id = world.deploy(module_bytecode!("box"))?;
+
+    fn create_snapshot(
+        world: &mut World,
+        id: ModuleId,
+        arg: i16,
+    ) -> Result<SnapshotId, Error> {
+        let _: Receipt<()> = world.transact(id, "set", arg)?;
+
+        let value = world.query::<_, Option<i16>>(id, "get", ())?;
+
+        assert_eq!(*value, Some(arg));
+
+        world.persist()
+    }
+
+    fn restore_snapshot(
+        world: &mut World,
+        id: ModuleId,
+        snapshot_id: &SnapshotId,
+        arg: i16,
+    ) -> Result<(), Error> {
+        world.restore(&snapshot_id)?;
+        let value = world.query::<_, Option<i16>>(id, "get", ())?;
+        let _: Receipt<()> = world.transact(id, "mem_snap", ())?;
+        assert_eq!(*value, Some(arg));
+        Ok(())
+    }
+
+    let mut snapshot_ids = Vec::new();
+    let random_i = vec![3, 1, 0, 4, 2];
+    for i in 0..random_i.len() {
+        snapshot_ids.push(create_snapshot(&mut world, id, i as i16)?);
+    }
+    for i in random_i {
+        restore_snapshot(&mut world, id, &snapshot_ids[i], (i) as i16)?;
+    }
+    Ok(())
+}
 
 #[test]
 pub fn snapshot_hash_excludes_argbuf() -> Result<(), Error> {
@@ -12,14 +56,10 @@ pub fn snapshot_hash_excludes_argbuf() -> Result<(), Error> {
     let id = world.deploy(module_bytecode!("box"))?;
 
     let snapshot_id1 = world.persist()?;
-    let _: Receipt<()> = world.transact(id, "mem_snap", ())?;
     let _: Receipt<()> = world.transact(id, "noop_query_with_arg", 0x22)?;
-    let _: Receipt<()> = world.transact(id, "mem_snap", ())?;
     let snapshot_id2 = world.persist()?;
     assert_ne!(snapshot_id1, snapshot_id2); // snapshot 1 has empty heap, not init-ed yet
-    let _: Receipt<()> = world.transact(id, "mem_snap", ())?;
     let _: Receipt<()> = world.transact(id, "noop_query_with_arg", 0x22)?;
-    let _: Receipt<()> = world.transact(id, "mem_snap", ())?;
     let snapshot_id3 = world.persist()?;
     assert_eq!(snapshot_id2, snapshot_id3);
